+++ libs/setenv/projectUrl/index.mjs
@@ -1,8 +1,8 @@
-import { existsSync } from "fs";
+import { existsSync, readFileSync } from "fs";
 
 export const projectUrl = () => {
   let baseUrl = new URL(`../../../`, new URL(import.meta.url));
-  while (!existsSync(new URL("environments", baseUrl))) {
+  while (!isBase(baseUrl)) {
     const parentUrl = new URL(`..`, baseUrl);
     if (parentUrl.toString() == baseUrl.toString()) {
       throw new Error("Could not find project root");
@@ -11,3 +11,14 @@ export const projectUrl = () => {
   }
   return baseUrl;
 };
+
+function isBase(url) {
+  const packagePath = new URL("package.json", url);
+  if (existsSync(packagePath)) {
+    // Check that it's the right one
+    const pkg = JSON.parse(readFileSync(packagePath, "utf8"));
+    // If we are somehow in our own folder, it's not the root.
+    return pkg.name !== "@thecointech/setenv";
+  }
+  return false;
+}