+++ tools/predeploy.ts
@@ -1,9 +1,8 @@
 import { getEnvVars } from "@thecointech/setenv";
 import { exit } from "process";
 import { readFileSync, writeFileSync, existsSync } from "fs";
-import { spawn } from 'child_process';
+import { spawn, exec as exec_cb } from 'child_process';
 import { promisify } from 'util';
-import { exec as exec_cb } from 'child_process';
 import { getSecret } from "@thecointech/secrets";
 import { join } from 'path';
 
@@ -103,6 +102,9 @@ export async function copyNpmTokenHere(folder: URL) {
   const noToken = readFileSync(new URL('.npmrc', import.meta.url), 'utf8');
   const withToken = noToken.replace('<tokenhere>', token);
   writeFileSync(new URL('.npmrc', folder), withToken);
+
+  console.log(`NPM Token: ${token.substring(0, 8)}... written to ${folder}`);
+
   return "NPM token copied here";
 }